restic: init repository if needed, by default
Add backup/init config directive (defaults to yes) to control whether
the repository should be initialized if it doesn't exist. Since we need
to attempt to connect to the repository to verify if init is needed,
remove the testconnect parameter so that the handler just tests the
connection every time it runs.