assume clean build environment, see DEBBUG-915222
This is part of a fix for reproducible builds where usrmerge may or may
not be in effect on the build system. We need to be able to pass the
BASH envvar to autotools so its able to correctly determine the correct
path for bash. The lines removed here were added to fix DEBBUG-346303
but should not be required on a clean build environment, where BASH
remains unset unless the current shell is actually BASH, and when it is,
should return the correct path to the bash executable and not /bin/sh.